% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_tokenz.R
\name{num_tokenz}
\alias{num_tokenz}
\title{How many tokens are in the strings?}
\usage{
num_tokenz(s)
}
\arguments{
\item{s}{vector of strings}
}
\value{
a number
}
\description{
How many tokens are in the strings?
}
\examples{
num_tokenz(c("a bad wolf", "stole my", "oranges"))
}
