% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_func.R
\name{trial_func}
\alias{trial_func}
\title{Detect h in strings}
\usage{
trial_func(a)
}
\arguments{
\item{a}{vector of strings}
}
\value{
boolean vector
}
\description{
Return boolean vector that tells whether strings contain h
}
\examples{
trial_func(c("hi","bye","jello holiday"))
}
